backend/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
openai==0.27.0
requests==2.28.1
python-dotenv==1.0.0
werkzeug==2.2.2